+++ beast-of-argh/CHANGELOG.md
@@ -1,3 +1,9 @@
+0.22.2
+
+* Enable milliseconds in syslog-ng for all log timestamps
+
+---
+
 0.22
 
 * Version bump for new base image