+++ haproxy-sql/CHANGELOG.md
@@ -1,3 +1,9 @@
+0.14.2
+
+* Enable milliseconds in syslog-ng for all log timestamps
+
+---
+
 0.14
 
 * Version bump for new base image