+++ mariadb/CHANGELOG.md
@@ -1,3 +1,9 @@
+4.10.2
+
+* Enable milliseconds in syslog-ng for all log timestamps
+
+---
+
 4.10
 
 * Version bump for new base image