+++ matrix-synapse/CHANGELOG.md
@@ -1,3 +1,9 @@
+2.10.2
+
+* Enable milliseconds in syslog-ng for all log timestamps
+
+---
+
 2.10
 
 * Version bump for new base image