+++ nextcloud-spreed-signalling/CHANGELOG.md
@@ -1,3 +1,9 @@
+0.3.2
+
+* Enable milliseconds in syslog-ng for all log timestamps
+
+---
+
 0.3
 
 * Version bump for new base image