+++ nginx-rsync-ssh/CHANGELOG.md
@@ -1,3 +1,9 @@
+0.21.2
+
+* Enable milliseconds in syslog-ng for all log timestamps
+
+---
+
 0.21
 
 * Version bump for new base image