+++ postgres-single/CHANGELOG.md
@@ -1,3 +1,9 @@
+0.12.2
+
+* Enable milliseconds in syslog-ng for all log timestamps
+
+---
+
 0.12
 
 * Version bump for new base image