+++ traumadrill/CHANGELOG.md
@@ -1,3 +1,9 @@
+1.22.2
+
+* Enable milliseconds in syslog-ng for all log timestamps
+
+---
+
 1.22
 
 * Version bump for new base image