+++ zincsearch/CHANGELOG.md
@@ -1,3 +1,9 @@
+0.4.2
+
+* Enable milliseconds in syslog-ng for all log timestamps
+
+---
+
 0.4
 
 * Version bump for new base image